+++ packages/terra-worklist-data-grid/tests/wdio/worklist-data-grid-spec.js
@@ -184,6 +184,18 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       Terra.validates.element('cell-3-1-focused', { selector: defaultSelector });
     });
+
+    it('hovers over selectable cell on an even row', () => {
+      browser.$$('tbody tr')[0].$$('td')[0].moveTo();
+      browser.pause(2000);
+      Terra.validates.element('hover-odd-row-cell-selection', { selector: defaultSelector });
+    });
+
+    it('hovers over selectable cell on an even row', () => {
+      browser.$$('tbody tr')[3].$$('td')[0].moveTo();
+      browser.pause(2000);
+      Terra.validates.element('hover-even-row-cell-selection', { selector: defaultSelector });
+    });
   });
 
   describe('With row selection', () => {
@@ -284,12 +296,24 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       Terra.validates.element('hover-selectable-cell', '#terra-worklist-data-grid-with-selections');
     });
 
-    it('hovers over non-selectable cell in worklist data grid', () => {
+    it('hovers over selectable cell in worklist data grid', () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/worklist-data-grid-with-selections');
 
-      browser.$$('td[class*="worklist-data-grid-cell"]:not([class*="selectable"])')[0].moveTo();
+      browser.$$('td[class*="worklist-data-grid-cell"][class*="selectable"]')[0].moveTo();
       browser.pause(2000);
-      Terra.validates.element('hover-non-selectable-cell', '#terra-worklist-data-grid-with-selections');
+      Terra.validates.element('hover-selectable-cell', '#terra-worklist-data-grid-with-selections');
+    });
+
+    it('hovers over selectable cell on an even row', () => {
+      browser.$$('tbody tr')[0].$$('td')[0].moveTo();
+      browser.pause(2000);
+      Terra.validates.element('hover-odd-row-selection', { selector: defaultSelector });
+    });
+
+    it('hovers over selectable cell on an even row', () => {
+      browser.$$('tbody tr')[3].$$('td')[0].moveTo();
+      browser.pause(2000);
+      Terra.validates.element('hover-even-row-selection', { selector: defaultSelector });
     });
 
     it('enables row selection mode by Shift+Space when turned off', () => {
@@ -542,6 +566,20 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
     const pinnedColumnsSelector = '#pinned-columns';
     const pinnedColumnsWithRowSelectionSelector = '#pinned-columns-with-row-selection';
 
+    it('validates that focused cells scroll into view when pinned columns exist', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/pinned-columns');
+
+      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight']);
+
+      const testScript = 'document.querySelector(\'div[class*="worklist-data-grid-container"\').scrollBy(450,0)';
+      browser.execute(testScript);
+
+      browser.keys(['ArrowRight']);
+
+      browser.pause(150);
+      Terra.validates.element('pinned-column-scroll-fix', { selector: pinnedColumnsSelector });
+    });
+
     it('adds pinned columns', () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/pinned-columns');
 
@@ -620,6 +658,20 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       navigateToCell(2, 1);
       moveCurrentPositionBy(6, 0);
       Terra.validates.element('sticky-header-scroll', { selector: stickyHeaderSelector });
+    });
+
+    it('validates that focused cells scroll into view when beneath sticky column header', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/worklist-data-grid-sticky-header');
+
+      browser.keys(['Tab']);
+
+      const testScript = 'document.querySelector(\'div[class*="worklist-data-grid-container"\').scrollBy(0,25)';
+      browser.execute(testScript);
+
+      browser.keys(['ArrowDown']);
+
+      browser.pause(250);
+      Terra.validates.element('sticky-header-scroll-fix', { selector: stickyHeaderSelector });
     });
   });
 });